requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
jinja2>=3.1.0
python-multipart>=0.0.9
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
requests>=2.31.0
apscheduler>=3.10.0
playwright>=1.41.0
openai>=1.30.0
anthropic>=0.30.0
python-telegram-bot>=21.0
pydantic>=2.6.0
pytz>=2024.1
